Clean up warnings under Visual C++
git-svn-id: svn://sporkhack.com/nh@302 0b9a868d-f130-0410-bc2a-adc7ae8015a6

+++ src/potion.c
@@ -121,7 +121,7 @@ int
 resist_reduce(amount,which)
 int amount, which;
 {
-	float tmp = 100 - how_resistant(which);
+	float tmp = (float)(100 - how_resistant(which));
 	tmp /= 100;
 	return (int)((float)amount * tmp);
 }
